Save table of contest to post json
