+++ components/Post/DynamicPostComponents/TableOfContents.js
@@ -1,66 +1,11 @@
 /* eslint-disable prefer-destructuring */
-import parse from 'html-react-parser';
-import React, { useEffect, useState } from 'react';
+import React from 'react';
 import getSlug from 'speakingurl';
 
 require('./TableOfContents.css');
 
 const TableOfContents = props => {
-  const [headings, setHeadings] = useState([]);
-
-  useEffect(() => {
-    const { htmlBody } = props;
-
-    let lastHeadingLevel;
-    let lastIsSubheading = false;
-    const newHeadings = [];
-    let hash1 = false;
-
-    const parseOptions = {
-      replace: ({ name, children, attribs }) => {
-        if (!attribs) return;
-
-        if (
-          (name === 'h1' ||
-            name === 'h2' ||
-            name === 'h3' ||
-            (!hash1 && name === 'h4')) &&
-          children &&
-          children.length > 0 &&
-          children[0].data &&
-          typeof children[0].data === 'string' &&
-          name[1]
-        ) {
-          if (name === 'h1') hash1 = true;
-          const level = Number(name[1]);
-          const newHeading = {
-            title: children[0].data,
-            subheadings: [],
-          };
-          if (
-            lastHeadingLevel &&
-            (level === lastHeadingLevel + 1 ||
-              (lastIsSubheading && level === lastHeadingLevel))
-          ) {
-            if (!lastIsSubheading || level === lastHeadingLevel)
-              newHeadings[newHeadings.length - 1].subheadings.push(newHeading);
-            else {
-              const subheads = newHeadings[newHeadings.length - 1].subheadings;
-              subheads[subheads.length - 1].subheadings.push(newHeading);
-            }
-            lastIsSubheading = true;
-          } else {
-            newHeadings.push(newHeading);
-            lastIsSubheading = false;
-          }
-          lastHeadingLevel = level;
-        }
-      },
-    };
-
-    parse(htmlBody, parseOptions);
-    setHeadings(newHeadings);
-  }, [props]);
+  const { headings } = props;
 
   return (
     <>
